refact: 예약 폼 및 API 타입 수정
- 예약 폼에서 카테고리 가격 및 시간을 표시하는 문구 수정
- 예약 시간 선택 로직에서 최대 시간을 12시간으로 변경
- 예약 관련 API 타입을 reservation.ts로 이동 및 수정
- 불필요한 console.log 제거 및 코드 정리
