+++ apps/src/features/reservation/ui/TimePickerModal.tsx
@@ -86,10 +86,9 @@ export const TimePickerModal = ({
                     </span>
                   </div>
                   <p className="text-sm break-keep">
-                    고객님의 공간({Math.floor(recommendedTime.area)}평)에
-                    최적화된 청소 시간은{' '}
-                    {Math.floor(recommendedTime.minutes / 60)}시간입니다. 실제
-                    현장 상황에 따라 추가 시간이 필요할 수 있습니다.
+                    고객님의 공간({Math.floor(recommendedTime.area)}평)에 
+                    최적화된 청소 시간은 {Math.floor(recommendedTime.minutes / 60)}시간입니다. 
+                    실제 현장 상황에 따라 추가 시간이 필요할 수 있습니다.
                   </p>
                 </div>
               )}
@@ -111,19 +110,20 @@ export const TimePickerModal = ({
                   <p className="text-3xl font-bold text-gray-800">
                     {selectedHours}시간
                   </p>
-                  <p className="text-xs text-gray-500 mt-1">
-                    기본 {standardHours}시간
+                  <p className="text-xs text-gray-500 mt-2">
+                    최소 {standardHours}시간, 
+                    최대 {12}시간
                   </p>
                 </div>
                 <button
                   onClick={() => onTimeChange(true)}
                   className={`w-12 h-12 rounded-full flex items-center justify-center text-2xl font-bold transition-all
                     ${
-                      selectedHours >= 8
+                      selectedHours >= 12
                         ? 'bg-gray-100 text-gray-400 cursor-not-allowed'
                         : 'bg-white border-2 border-cyan-500 text-cyan-500 active:bg-cyan-50'
                     }`}
-                  disabled={selectedHours >= 8}
+                  disabled={selectedHours >= 12}
                 >
                   +
                 </button>
